requirements.txt
python-telegram-bot
requests
schedule